# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Trading Platform API
  version: 1.0.0
  description: |
    Complete REST API for the Trading Backtesting Platform.
    Provides endpoints for authentication, backtesting sessions, trade journaling,
    analytics, playbooks, broker integrations, and subscription management.
  contact:
    name: Trading Platform Support
    email: support@tradingplatform.com
    url: https://docs.tradingplatform.com
  license:
    name: Proprietary
    url: https://tradingplatform.com/terms
servers:
  - url: https://api.tradingplatform.com/v1
    description: Production
  - url: https://staging-api.tradingplatform.com/v1
    description: Staging
  - url: http://localhost:3000
    description: Local Development
...
paths:
...
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh an access token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /users/me:
    get:
      operationId: getProfile
      summary: Get current user profile
      tags: [Users]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
    patch:
      operationId: updateProfile
      summary: Update current user profile
      tags: [Users]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /backtesting/sessions:
    get:
      operationId: listSessions
      summary: List backtesting sessions
      tags: [Backtesting]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/SessionStatus'
      responses:
        '200':
          description: Paginated session list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSessions'
    post:
      operationId: createSession
      summary: Create a new backtesting session
      tags: [Backtesting]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSessionRequest'
            example:
              sessionName: My First Backtest
              instrument: AAPL
              assetClass: stock
              startingBalance: 10000
              startDate: '2024-01-01'
              endDate: '2024-12-31'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestingSession'
  /backtesting/sessions/{sessionId}:
    parameters:
      - $ref: '#/components/parameters/SessionId'
    get:
      operationId: getSession
      summary: Get session details
      tags: [Backtesting]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacktestingSession'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateSession
      summary: Update session (e.g. playback speed, status)
      tags: [Backtesting]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionRequest'
      responses:
        '200':
          description: Session updated
    delete:
      operationId: deleteSession
      summary: Delete a session
      tags: [Backtesting]
      security: [{ BearerAuth: [] }]
      responses:
        '204':
          description: Session deleted
...
  /trades:
    get:
      operationId: listTrades
      summary: List journal trades
      tags: [Trade Journal]
      security: [{ BearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: instrument
          in: query
          schema:
            type: string
        - name: direction
          in: query
          schema:
            $ref: '#/components/schemas/TradeDirection'
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Paginated trades
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTrades'
    post:
      operationId: createTrade
      summary: Log a new trade
      tags: [Trade Journal]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJournalTradeRequest'
      responses:
        '201':
          description: Trade logged
  /trades/{tradeId}:
    parameters:
      - name: tradeId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getTrade
      summary: Get trade details
      tags: [Trade Journal]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Trade details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
    patch:
      operationId: updateTrade
      summary: Update a trade
      tags: [Trade Journal]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTradeRequest'
      responses:
        '200':
          description: Trade updated
    delete:
      operationId: deleteTrade
      summary: Delete a trade
      tags: [Trade Journal]
      security: [{ BearerAuth: [] }]
      responses:
        '204':
          description: Trade deleted
...
  /playbooks:
    get:
      operationId: listPlaybooks
      summary: List playbooks
      tags: [Playbooks]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Playbook list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Playbook'
    post:
      operationId: createPlaybook
      summary: Create a playbook
      tags: [Playbooks]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlaybookRequest'
      responses:
        '201':
          description: Playbook created
  /playbooks/{playbookId}:
    parameters:
      - name: playbookId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getPlaybook
      summary: Get playbook details with performance
      tags: [Playbooks]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Playbook with performance metrics
    patch:
      operationId: updatePlaybook
      summary: Update a playbook
      tags: [Playbooks]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlaybookRequest'
      responses:
        '200':
          description: Playbook updated
    delete:
      operationId: deletePlaybook
      summary: Delete a playbook
      tags: [Playbooks]
      security: [{ BearerAuth: [] }]
      responses:
        '204':
          description: Playbook deleted

  # ── Analytics ───────────────────────────────
  /analytics/overview:
    get:
      operationId: getOverview
      summary: Get performance overview
      tags: [Analytics]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y, all]
      responses:
        '200':
          description: Overview metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsOverview'
  /analytics/equity-curve:
    get:
      operationId: getEquityCurve
      summary: Get equity curve data points
      tags: [Analytics]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Equity curve
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                      format: date
                    balance:
                      type: number
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    SubscriptionTier:
      type: string
      enum: [free, pro, team, enterprise]
    SessionStatus:
      type: string
      enum: [active, paused, completed, archived]
    TradeDirection:
      type: string
      enum: [long, short]
    AssetClass:
      type: string
      enum: [stock, forex, crypto, futures, options, index]
    TradeSource:
      type: string
      enum: [backtest, live, manual, imported]

    # ── Auth ──
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string

    # ── User ──
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        subscriptionTier:
          $ref: '#/components/schemas/SubscriptionTier'
        isEmailVerified:
          type: boolean
        twoFactorEnabled:
          type: boolean
        preferences:
          type: object
        createdAt:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string

    # ── Backtesting ──
    BacktestingSession:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sessionName:
          type: string
        instrument:
          type: string
        assetClass:
          $ref: '#/components/schemas/AssetClass'
        startingBalance:
          type: number
        currentBalance:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        playbackSpeed:
          type: integer
        status:
          $ref: '#/components/schemas/SessionStatus'
        createdAt:
          type: string
          format: date-time
    CreateSessionRequest:
      type: object
      required: [sessionName, instrument, assetClass, startingBalance, startDate, endDate]
      properties:
        sessionName:
          type: string
        instrument:
          type: string
        assetClass:
          $ref: '#/components/schemas/AssetClass'
        startingBalance:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        playbookId:
          type: string
          format: uuid
    UpdateSessionRequest:
      type: object
      properties:
        playbackSpeed:
          type: integer
        status:
          $ref: '#/components/schemas/SessionStatus'
    PaginatedSessions:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BacktestingSession'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Trades ──
    BacktestingTrade:
      type: object
      properties:
        id:
          type: string
          format: uuid
        direction:
          $ref: '#/components/schemas/TradeDirection'
        entryPrice:
          type: number
        exitPrice:
          type: number
        quantity:
          type: number
        stopLoss:
          type: number
        takeProfit:
          type: number
        pnlNet:
          type: number
        entryDatetime:
          type: string
          format: date-time
        exitDatetime:
          type: string
          format: date-time
...
    Trade:
      allOf:
        - $ref: '#/components/schemas/BacktestingTrade'
        - type: object
          properties:
            source:
              $ref: '#/components/schemas/TradeSource'
            instrument:
              type: string
            assetClass:
              $ref: '#/components/schemas/AssetClass'
            fees:
              type: number
            commission:
              type: number
            pnlPercentage:
              type: number
            riskRewardRatio:
              type: number
            tradeDurationMinutes:
              type: integer
            notes:
              type: string
            tags:
              type: array
              items:
                type: string
    CreateJournalTradeRequest:
      type: object
      required: [instrument, assetClass, direction, entryPrice, quantity, entryDatetime]
      properties:
        instrument:
          type: string
        assetClass:
          $ref: '#/components/schemas/AssetClass'
        direction:
          $ref: '#/components/schemas/TradeDirection'
        entryPrice:
          type: number
        exitPrice:
          type: number
        quantity:
          type: number
        entryDatetime:
          type: string
          format: date-time
        exitDatetime:
          type: string
          format: date-time
        stopLoss:
          type: number
        takeProfit:
          type: number
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
    UpdateTradeRequest:
      type: object
      properties:
        exitPrice:
          type: number
        exitDatetime:
          type: string
          format: date-time
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
    PaginatedTrades:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Trade'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Playbook:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        entryRules:
          type: array
          items:
            type: object
        exitRules:
          type: array
          items:
            type: object
        riskParameters:
          type: object
        instruments:
          type: array
          items:
            type: string
        timeframes:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
        version:
          type: integer
        performance:
          $ref: '#/components/schemas/PlaybookPerformance'
    PlaybookPerformance:
      type: object
      properties:
        totalTrades:
          type: integer
        winningTrades:
          type: integer
        losingTrades:
          type: integer
        winRate:
          type: number
        avgPnl:
          type: number
        profitFactor:
          type: number
        maxDrawdown:
          type: number
    CreatePlaybookRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        entryRules:
          type: array
          items:
            type: object
        exitRules:
          type: array
          items:
            type: object
        riskParameters:
          type: object
        instruments:
          type: array
          items:
            type: string
        timeframes:
          type: array
          items:
            type: string
    UpdatePlaybookRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        entryRules:
          type: array
          items:
            type: object
        exitRules:
          type: array
          items:
            type: object
        riskParameters:
          type: object

    # ── Analytics ──
    AnalyticsOverview:
      type: object
      properties:
        totalTrades:
          type: integer
        winRate:
          type: number
        profitFactor:
          type: number
        totalPnl:
          type: number
        avgWin:
          type: number
        avgLoss:
          type: number
        largestWin:
          type: number
        largestLoss:
          type: number
        maxDrawdown:
          type: number
        sharpeRatio:
          type: number
        currentStreak:
          type: integer

    # ── Broker ──
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
...